fix: cast cos/sin to x.dtype in apply_rotary — fix CUDA graph bf16 capture
RotaryEmbedding.forward() returns float32 cos/sin (positions.float() path).
When q/k are bfloat16, the element-wise multiplication caused dtype promotion
to float32, which then triggered 'float != bfloat16' on downstream matmuls
during CUDA graph capture.

+++ vllm_i64/layers/rotary.py
@@ -67,8 +67,8 @@ def apply_rotary(x: torch.Tensor, cos: torch.Tensor, sin: torch.Tensor) -> torch
     d = x.shape[-1] // 2
     x1, x2 = x[..., :d], x[..., d:]
     # cos/sin are (batch, head_dim) = (batch, 2*d), take first half
-    cos = cos[..., :d].unsqueeze(1)  # (batch, 1, d)
-    sin = sin[..., :d].unsqueeze(1)  # (batch, 1, d)
+    cos = cos[..., :d].unsqueeze(1).to(x.dtype)  # (batch, 1, d)
+    sin = sin[..., :d].unsqueeze(1).to(x.dtype)  # (batch, 1, d)
     return torch.cat([x1 * cos - x2 * sin, x2 * cos + x1 * sin], dim=-1)
